requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.5
nose==1.3.0
oauthlib==2.0.6
py-trello==0.10.0
python-dateutil==2.4.2
pytz==2018.5
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
urllib3==1.21.1
